feat(e2e): refactors queryserving e2e location
